requirements.txt
psycopg2-binary
gunicorn==20.1.0
requests==2.25.1
bcrypt
